requirements.txt
openai>=1.0.0
pyyaml>=6.0
pytest>=7.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.28.0

# LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# RAG - Local Knowledge Base
pypdf>=4.0.0
pdfplumber>=0.11.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
